Add proper B0505 reference design
