+++ src/faebryk/library/B0505S_1WR3.py
@@ -12,7 +12,7 @@ from faebryk.libs.units import P  # noqa: F401
 logger = logging.getLogger(__name__)
 
 
-class _B0505S_1WR3(Module):
+class B0505S_1WR3(Module):
     """
     Isolated 5V DC to 5V DC converter.
     R suffix is for shortcircuit protection
@@ -64,36 +64,3 @@ class _B0505S_1WR3(Module):
     def __preinit__(self):
         self.power_in.voltage.constrain_subset(L.Range(4.3 * P.V, 9 * P.V))
         self.power_out.voltage.constrain_superset(L.Range.from_center_rel(5 * P.V, 0.1))
-
-
-class B0505S_1WR3_ReferenceDesign(Module):
-    ic: _B0505S_1WR3
-
-    power_in: F.ElectricPower
-    power_out: F.ElectricPower
-
-    ferrite_in: F.Inductor
-    ferrite_out: F.Inductor
-
-    def __preinit__(self):
-        self.power_in.voltage.alias_is(self.ic.power_in.voltage)
-        self.power_out.voltage.alias_is(self.ic.power_out.voltage)
-
-        self.ferrite_in.inductance.constrain_subset(
-            L.Range.from_center_rel(4.7 * P.uH, 0.1)
-        )
-        self.ferrite_out.inductance.constrain_subset(
-            L.Range.from_center_rel(4.7 * P.uH, 0.1)
-        )
-        self.power_in.decoupled.decouple(owner=self).capacitance.constrain_subset(
-            L.Range.from_center_rel(4.7 * P.uF, 0.2)
-        )
-        self.power_out.decoupled.decouple(owner=self).capacitance.constrain_subset(
-            L.Range.from_center_rel(10 * P.uF, 0.2)
-        )
-
-        self.power_in.hv.connect_via(self.ferrite_in, self.ic.power_in.hv)
-        self.power_in.lv.connect(self.ic.power_in.lv)
-
-        self.power_out.hv.connect_via(self.ferrite_out, self.ic.power_out.hv)
-        self.power_out.lv.connect(self.ic.power_out.lv)
